Refactored the code so that the LinearSlideTest class now inherits from TeleopTemplate. Moved all Teleop to a dedicated folder.
