+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/src/DrivePrograms/Teleop/JavaDriveProgram.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.src.DrivePrograms;
+package org.firstinspires.ftc.teamcode.src.DrivePrograms.Teleop;
 
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
